[proliferate] clean cloud credential domain rules
