+++ server/proliferate/server/ai_magic/errors.py
@@ -5,4 +5,3 @@ from proliferate.errors import ProliferateError
 
 class AiMagicError(ProliferateError):
     """Raised when an AI magic request fails with a client-facing error."""
-
