+++ server/proliferate/server/cloud/credentials/domain/__init__.py
@@ -1,0 +1,1 @@
+"""Pure cloud credential domain helpers."""
