+++ server/proliferate/server/cloud/credentials/domain/types.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from typing import Literal
+
+CloudCredentialAuthMode = Literal["env", "file"]
