+++ server/proliferate/server/cloud/credentials/models.py
@@ -1,26 +1,15 @@
-"""Request schemas and domain types for cloud credentials."""
+"""Request and response schemas for cloud credentials."""
 
 from __future__ import annotations
 
-from collections.abc import Sequence
-from dataclasses import dataclass
-from datetime import datetime
 from typing import Literal
 
 from pydantic import BaseModel, Field
 
-from proliferate.constants.cloud import (
-    SUPPORTED_CLOUD_AGENTS,
+from proliferate.server.cloud.credentials.domain.status import (
+    CredentialStatusRecord,
 )
-from proliferate.constants.cloud import CloudAgentKind as CloudAgentKind
-
-CloudCredentialAuthMode = Literal["env", "file"]
-
-_DEFAULT_AUTH_MODES: dict[str, CloudCredentialAuthMode] = {
-    "claude": "env",
-    "codex": "file",
-    "gemini": "env",
-}
+from proliferate.server.cloud.credentials.domain.types import CloudCredentialAuthMode
 
 
 class SyncClaudeEnvCredentialRequest(BaseModel):
@@ -76,21 +65,6 @@ SyncCloudCredentialRequest = (
 )
 
 
-# ---------------------------------------------------------------------------
-# Credential status domain types
-# ---------------------------------------------------------------------------
-
-
-@dataclass(frozen=True)
-class CredentialStatusRecord:
-    provider: CloudAgentKind
-    auth_mode: CloudCredentialAuthMode
-    supported: bool
-    local_detected: bool
-    synced: bool
-    last_synced_at: str | None
-
-
 class CredentialStatus(BaseModel):
     provider: str
     auth_mode: CloudCredentialAuthMode = Field(serialization_alias="authMode")
@@ -105,52 +79,6 @@ class CloudCredentialMutationResponse(BaseModel):
     changed: bool
 
 
-def _to_iso(value: object) -> str | None:
-    if value is None:
-        return None
-    if isinstance(value, datetime):
-        return value.isoformat()
-    return str(value)
-
-
-def build_credential_statuses(
-    records: Sequence[object],
-) -> list[CredentialStatusRecord]:
-    """Build status records from a sequence of ORM credential rows.
-
-    Each element is expected to have ``provider``, ``auth_mode``,
-    ``revoked_at``, and ``last_synced_at`` attributes (i.e. the
-    ``CloudCredential`` ORM model).
-    """
-    by_provider: dict[str, object] = {}
-    for record in records:
-        provider = getattr(record, "provider", None)
-        if provider in SUPPORTED_CLOUD_AGENTS and getattr(record, "revoked_at", None) is None:
-            by_provider[provider] = record
-
-    statuses: list[CredentialStatusRecord] = []
-    for provider in SUPPORTED_CLOUD_AGENTS:
-        record = by_provider.get(provider)
-        raw_auth_mode = (
-            getattr(record, "auth_mode", None) if record else None
-        ) or _DEFAULT_AUTH_MODES.get(provider, "env")
-        auth_mode: CloudCredentialAuthMode = (
-            raw_auth_mode if raw_auth_mode in ("env", "file") else "env"
-        )
-        last_synced_at = _to_iso(getattr(record, "last_synced_at", None)) if record else None
-        statuses.append(
-            CredentialStatusRecord(
-                provider=provider,
-                auth_mode=auth_mode,
-                supported=True,
-                local_detected=False,
-                synced=record is not None,
-                last_synced_at=last_synced_at,
-            )
-        )
-    return statuses
-
-
 def credential_status_payload(status: CredentialStatusRecord) -> CredentialStatus:
     """Serialize a *CredentialStatusRecord* to its wire-format Pydantic model."""
     return CredentialStatus(
@@ -161,13 +89,3 @@ def credential_status_payload(status: CredentialStatusRecord) -> CredentialStatu
         synced=status.synced,
         last_synced_at=status.last_synced_at,
     )
-
-
-def allowed_agent_kinds() -> list[str]:
-    """Return the list of provider names that the platform supports."""
-    return list(SUPPORTED_CLOUD_AGENTS)
-
-
-def ready_agent_kinds(statuses: Sequence[CredentialStatusRecord]) -> list[str]:
-    """Return provider names that have a synced credential."""
-    return [status.provider for status in statuses if status.synced]
